Int to long for podioId in Employee and updated the podio nuget

+++ API/ShiftPlanning.DTOs/Employee/EmployeeDTO.cs
@@ -12,6 +12,6 @@
         public int? CheckInCount { get; set; }
         public int WantShifts { get; set; }
         public string[] Roles { get; set; }
-        public int PodioId { get; set; }
+        public long PodioId { get; set; }
     }
 }
